issue: 2246994 Set proper FIN/RST flags for split segments
When a segment with FIN flag is split into multiple segments, the
leftmost segment retains all its flags including FIN. This leads to
corrupted TCP stream.

To fix this, move FIN/RST flags to the rightmost among the split
segments. Currently, VMA doesn't support RST flag in the pending
segments, however, it may be changed in the future.

Signed-off-by: Dmytro Podgornyi <dmytrop@mellanox.com>